refactor(ai): decompose moderation module to match schedule_guard architecture
- Split monolithic moderation/mod.rs into separate components:
  * dto.rs: ModerationResult and ModerationOverrides structs
  * overrides.rs: build_override_section utility function
  * moderation_service.rs: core ModerationService implementation
  * mod.rs: module declarations and re-exports

- Inject ModerationService into BotDependencies for consistency
- Replace ad-hoc ModerationService::new() calls with dependency injection
- Add Clone derive to ModerationService for async sharing
- Maintain public API compatibility via re-exports

This creates architectural consistency between moderation/ and schedule_guard/
modules, improves maintainability through separation of concerns, and
enables better testing of individual components.
