+++ quark_bot/src/main.rs
@@ -53,6 +53,7 @@ use crate::bot::handler_tree::handler_tree;
 use crate::scheduled_prompts::handler::bootstrap_scheduled_prompts;
 use tokio_cron_scheduler::JobScheduler;
 use crate::ai::schedule_guard::schedule_guard_service::ScheduleGuardService;
+use crate::ai::moderation::ModerationService;
 
 #[tokio::main(flavor = "multi_thread", worker_threads = 4)]
 async fn main() {
@@ -119,6 +120,8 @@ async fn main() {
     let ai = AI::new(openai_api_key.clone(), google_cloud);
     let schedule_guard = ScheduleGuardService::new(openai_api_key.clone())
         .expect("Failed to create ScheduleGuardService");
+    let moderation = ModerationService::new(openai_api_key.clone())
+        .expect("Failed to create ModerationService");
 
     let user_convos = UserConversations::new(&db).unwrap();
     let user_model_prefs = UserModelPreferences::new(&db).unwrap();
@@ -226,6 +229,7 @@ async fn main() {
         payment: payment.clone(),
         default_payment_prefs,
         schedule_guard: schedule_guard,
+        moderation: moderation,
     };
 
     // Bootstrap user-defined schedules (load and register)
